requirements.txt
litellm
pydantic
tqdm
python-dotenv
